+++ frontend/pages/review/[reviewId]/4-submit.tsx
@@ -1,38 +1,20 @@
-import Card from '@mui/material/Card';
-import CardContent from '@mui/material/CardContent';
-import CardHeader from '@mui/material/CardHeader';
-import Container from '@mui/material/Container';
-import Typography from '@mui/material/Typography';
-import Grid from '@mui/material/Grid';
-import React, { ReactNode, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useFormik } from 'formik';
 import * as Yup from 'yup';
-import { API_ENDPOINT, CourseEvaluationDetailEntry, ReviewListEntry } from 'utils/api';
+import { API_ENDPOINT } from 'utils/api';
 import TextField from '@mui/material/TextField';
 import { useSWRConfig } from 'swr';
 import Alert from '@mui/material/Alert';
 import useAuthenticatedAPIClient from '@/components/hooks/useAuthenticatedAPIClient';
-import EOCAccordionWithModal from '@/components/Reviewer/Assessment/EOCAccordionWithModal';
-import SimplifiedDocumentCard from '@/components/Reviewer/Submit/SimplifiedDocumentCard';
-import { getReviewStepsWithState, ReviewStepWithStateType } from '@/components/utils/reviews';
+import { getReviewStepsWithState } from '@/components/utils/reviews';
 import BodyCard from '@/components/utils/BodyCard';
 import ReviewProgress from '@/components/Reviewer/ReviewProgress';
 import ReviewerBottomNavigation from '@/components/Reviewer/ReviewerBottomNavigation';
 import AboutStepCard from '@/components/Reviewer/AboutStepCard';
 import useCourseReview from '@/components/hooks/useCourseReview';
 import useCourseEvaluation from '@/components/hooks/useCourseEvaluation';
-
-const StepWrapper = (props: { children: ReactNode; cardTitle: string }) => {
-  const { children, cardTitle } = props;
-  return (
-    <Container maxWidth="xl" sx={{ mt: 2, mb: 2 }}>
-      <Card>
-        <CardHeader title={cardTitle} />
-        <CardContent>{children}</CardContent>
-      </Card>
-    </Container>
-  );
-};
+import StepWrapper from '@/components/Reviewer/Submit/StepWrapper';
+import ReviewSummarySubmissionContent from '@/components/Reviewer/Submit/ReviewSummarySubmissionContent';
 
 const Submit = () => {
   const { courseReview } = useCourseReview();
@@ -101,7 +83,7 @@ const Submit = () => {
         allSteps={allSteps}
         courseReview={courseReview}
         courseEvaluation={courseEvaluation}
-        displayLastStep={false}
+        isReadOnly={false}
       />
       <StepWrapper cardTitle={`Step 4 - ${allSteps[3].stepName}`}>
         <TextField
@@ -137,60 +119,3 @@ const Submit = () => {
 };
 
 export default Submit;
-
-interface ReviewSummarySubmissionContentType {
-  allSteps: ReviewStepWithStateType[];
-  courseReview: ReviewListEntry;
-  courseEvaluation: CourseEvaluationDetailEntry;
-  isReadOnly?: boolean;
-}
-const ReviewSummarySubmissionContent = (props: ReviewSummarySubmissionContentType) => {
-  const { allSteps, courseReview, courseEvaluation, isReadOnly } = props;
-
-  return (
-    <>
-      <StepWrapper cardTitle={`Step 1 - ${allSteps[0].stepName}`}>
-        {courseReview.eoc_date_viewed ? (
-          <Typography color="secondary.main">
-            You have read and confirmed that you have understood the elements of competencies.
-          </Typography>
-        ) : (
-          <Typography color="error.main">
-            You have not yet read and confirmed that you have understood the elements of
-            competencies.
-          </Typography>
-        )}
-      </StepWrapper>
-      <StepWrapper cardTitle={`Step 2 - ${allSteps[1].stepName}`}>
-        <Grid container spacing={2}>
-          {courseEvaluation.documents.map((document) => (
-            <Grid item sm={12} md={4} key={document.id}>
-              <SimplifiedDocumentCard document={document} review={courseReview} />
-            </Grid>
-          ))}
-        </Grid>
-      </StepWrapper>
-      <StepWrapper cardTitle={`Step 3 - ${allSteps[2].stepName}`}>
-        <EOCAccordionWithModal courseEvaluation={courseEvaluation} courseReview={courseReview} />
-      </StepWrapper>
-      {!isReadOnly && (
-        <StepWrapper cardTitle={`Step 4 - ${allSteps[3].stepName}`}>
-          <TextField
-            margin="dense"
-            id="final_comment"
-            label="Final Comment"
-            fullWidth
-            variant="outlined"
-            multiline
-            disabled
-            value={courseReview.final_comment}
-          />
-        </StepWrapper>
-      )}
-    </>
-  );
-};
-
-ReviewSummarySubmissionContent.defaultProps = {
-  isReadOnly: true,
-};
